requirements.txt
fastapi
uvicorn
pdfplumber
pytesseract
Pillow
python-docx
docx2txt
spacy
python-multipart
easyocr